requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alabaster==0.7.12
annotated-types==0.7.0
anyio==3.7.1
apturl==0.5.2
arabic-reshaper==3.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
assemblyai==0.41.1
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
Automat==24.8.1
babel==2.17.0
bcrypt==3.2.0
beartype==0.21.0
beautifulsoup4==4.12.3
bleach==6.2.0
blessed==1.20.0
blinker==1.9.0
blis==1.3.0
Brlapi==0.8.3
cachetools==5.5.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
confection==0.1.5
constantly==23.10.4
contexttimer==0.3.3
contourpy==1.3.1
cryptography==41.0.5
cssselect==1.2.0
cssselect2==0.7.0
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.1+ubuntu0.2
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.7.0
docutils==0.17.1
duplicity==0.8.21
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastai==2.8.2
fastapi==0.103.2
fastcore==1.8.5
fastdownload==0.0.7
fasteners==0.14.1
fastjsonschema==2.21.1
fastprogress==1.0.3
fasttransform==0.0.2
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.16.1
Flask-Cors==5.0.0
fonttools==4.55.6
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.6.1
future==0.18.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.2.1
grpcio==1.68.1
grpcio-status==1.68.1
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
imagesize==1.3.0
importlib-metadata==4.6.4
incremental==24.7.2
iniconfig==1.1.1
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==23.5.0
kiwisolver==1.4.8
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.55
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.3.34
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.33
language-selector==0.1
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
loguru==0.7.3
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
meson==0.61.2
mistune==3.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
murmurhash==1.0.13
mypy_extensions==1.1.0
natsort==8.0.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
notebook_shim==0.2.4
numpy==2.2.5
oauthlib==3.2.0
olefile==0.46
openai==1.76.0
orjson==3.10.16
ormsgpack==1.9.1
oscrypto==1.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
paramiko==2.9.3
parsel==1.9.1
parso==0.8.4
pexpect==4.8.0
pillow==11.3.0
pinecone-client==2.2.4
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
pluggy==0.13.0
plum-dispatch==2.5.7
ply==3.11
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.10.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyDispatcher==2.0.7
Pygments==2.19.2
PyGObject==3.42.1
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.10.1
pymacaroons==0.13.0
pymongo==4.10.1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==2.4.7
pypdf==5.0.1
pyRFC3339==1.1
PySocks==1.7.1
pytest==6.2.5
python-apt==2.4.0+ubuntu2
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.2
pyzmq==27.0.0
qrcode==8.0
queuelib==1.7.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.2.5
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
roman==3.3
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
Scrapy==2.11.2
SecretStorage==3.3.1
selenium==4.25.0
Send2Trash==1.8.3
sentry-sdk==2.32.0
service-identity==24.1.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==4.3.2
sphinx-copybutton==0.4.0
sphinx-issues==1.2.0
sphinx-rtd-theme==1.0.0
SQLAlchemy==2.0.40
sqlparse==0.5.1
srsly==2.5.1
sse-starlette==1.6.5
stack-data==0.6.3
starlette==0.27.0
svglib==1.5.1
sympy==1.13.3
systemd-python==234
tenacity==9.1.2
tensorboardX==2.6.4
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.3.0
tldextract==5.1.2
toml==0.10.2
tomli==2.0.2
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.7.0
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.4.0
usb-creator==0.3.7
uuid==1.30
uvicorn==0.22.0
w3lib==2.2.1
wadllib==1.3.6
wandb==0.21.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
win32_setctime==1.2.0
wrapt==1.17.2
wsproto==1.2.0
wxPython==4.0.7
xdg==5
xhtml2pdf==0.2.16
xkit==0.0.0
xxhash==3.5.0
yarl==1.20.0
yt-dlp==2025.6.30
zipp==1.0.0
zope.interface==7.1.1
zstandard==0.23.0
